Support postings list compression using varint #19 (#24)
* implementation of PostingListWriter for varint postings #19

* add IPostingsListBuilder

to represent PostingsList Builder functionality

* fixed issue when PostingLists reused between indices

* implement VarintPostingsListBuilder

to support Varint listings posts

* minor fix

+++ src/InformationRetrieval/Indexing/External/BlockedExternalBuildableIndex.cs
@@ -198,7 +198,10 @@ namespace InformationRetrieval.Indexing.External
                 }
                 indices.Clear();
 
-                fileSystem.Directory.Delete(basePath, recursive: true);
+                if (fileSystem.Directory.Exists(basePath))
+                {
+                    fileSystem.Directory.Delete(basePath, recursive: true);
+                }
             }
 
             public void Dispose()
